+++ src/Modules/Profile/components/Profile.vue
@@ -103,6 +103,7 @@
           <Wall
             target="foodsaver"
             :target-id="profileInfos.fsId"
+            :limit="200"
           />
         </b-col>
       </b-row>