+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Removed application wall #321 !3436 @AntonBallmaier
 - Removed store state COOPERATION_STARTING #414 !3410 !3444 !3445 !3448 @chriswalg
 - Added the WhatsApp link to the social icons in the footer #2035 !3484 @alex.simm
+- The profile wall now shows up to 200 posts #2055 !3505 @AntonBallmaier
 
 ## Bugfixes
 - Changed the headline for group applications #1916 !3353 @ToLei